+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Moved aggregate out of full table query for label list [#5584](https://github.com/raster-foundry/raster-foundry/pull/5584)
 
+### Changed
+- Session complete now deactivates previous labels [#5583](https://github.com/raster-foundry/raster-foundry/pull/5583)
+
 ## [1.64.1] - 2021-05-18
 ### Fixed
 - Ensured task actions are updated in single request for labels and task status [#5581](https://github.com/raster-foundry/raster-foundry/pull/5581)